+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.1.1.0
+
+* Add convenience functions `bytestringBody`, `lazyBytestringBody`, `utf8TextBody`, `utf8LazyTextBody`
+* Add support for Brotli HTTP compression
+
 ## 0.1.0.0
 
 * First version. Released on an unsuspecting world.